+++ src/lib/routes/client-api/metrics.ts
@@ -2,12 +2,12 @@ import { Response } from 'express';
 import Controller from '../controller';
 import { IUnleashServices } from '../../types';
 import { IUnleashConfig } from '../../types/option';
-import ClientInstanceService from '../../services/client-metrics/client-instance-service';
+import ClientInstanceService from '../../services/client-metrics/instance-service';
 import { Logger } from '../../logger';
 import { IAuthRequest } from '../unleash-types';
 import ApiUser from '../../types/api-user';
 import { ALL } from '../../types/models/api-token';
-import ClientMetricsServiceV2 from '../../services/client-metrics/client-metrics-service-v2';
+import ClientMetricsServiceV2 from '../../services/client-metrics/metrics-service-v2';
 import { User } from '../../server-impl';
 import { IClientApp } from '../../types/model';
 import { NONE } from '../../types/permissions';
